moved AgGridTable to its own file
